fix: break suspension cache type cycle
